+++ app/lib/l10n/gen/ouds_flutter_app_localizations_en.dart
@@ -131,4 +131,10 @@ class AppLocalizationsEn extends AppLocalizations {
 
   @override
   String get app_about_changelog_label => 'Changelog';
+
+  @override
+  String get app_tokens_color_label => 'Color';
+
+  @override
+  String get app_tokens_color_description_text => 'Colour reinforces our brand identity and ensures consistency across all product experiences. The semantic tokens described are the onse you should use when building a mobile app.';
 }