+++ app/lib/ui/tokens/tokens.dart
@@ -16,6 +16,8 @@ import 'package:ouds_flutter_demo/ui/tokens/elevation/elevation_screen.dart';
 import 'package:ouds_flutter_demo/ui/tokens/token_entities.dart';
 import 'package:ouds_flutter_demo/ui/utilities/adaptive_image_helper.dart';
 
+import 'color/color_screen.dart';
+
 List<Token> tokens(BuildContext context) {
   return [
     Token(
@@ -23,6 +25,12 @@ List<Token> tokens(BuildContext context) {
       AdaptiveImageHelper.getImage(context, 'assets/il_elevation.png'),
       AppLocalizations.of(context)!.app_tokens_elevation_description_text,
       const ElevationScreen(),
+    ),
+    Token(
+      context.l10n.app_tokens_color_label,
+      AdaptiveImageHelper.getImage(context, 'assets/il_color.png'),
+      AppLocalizations.of(context)!.app_tokens_color_description_text,
+      const ColorScreen(),
     ),
   ];
 }